+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Planned]
 
 ## [Upcoming]
+### Changed
+Do not check explicitly the argument modified in checks.extends.
 
 ## [0.9.2] - 2025-05-27
 ### Changed